pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-backend"
version = "1.0.0"
description = "OCR backend: image → searchable PDF, with English, Hindi and Marathi support."
requires-python = ">=3.14"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "easyocr>=1.7",
    "opencv-python-headless>=4.10",
    "Pillow>=10.4",
    "reportlab>=4.2",
    "pydantic-settings>=2.5",
    "structlog>=24.4",
    "python-magic>=0.4",
    "python-multipart>=0.0.12",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
